+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Release notes
 
+### 0.13.3 -- May 28, 2021
+* Bugfix - Dependencies not properly loaded on populate. (#902) PR #919
+
 ### 0.13.2 -- May 7, 2021
 * Update `setuptools_certificate` dependency to new name `otumat`
 * Bugfix - Explicit calls to `dj.Connection` throw error due to missing `host_input` (#895) PR #907